@@ -5,7 +5,7 @@ ext.moduleName = 'api'
 
 apply plugin: 'java-library'
 
-sourceCompatibility = 1.10
+sourceCompatibility = 1.11
 
 repositories {
     mavenCentral()
